requirements.txt
pyserial==3.13.9
matplotlib==3.10.8
pynput==1.8.1